pyproject.toml
[tool.poetry]
name = "eivelia_bot"
version = "0.1.0"
description = "eivelia_bot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = "^2.0.0-beta.1"
nonebot-adapter-onebot = "^2.1.5"
gmqtt = "^0.6.11"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = []
plugin_dirs = ["eivelia_bot/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
